# api/openapi.yaml
openapi: 3.0.3
servers:
  - url: /
    description: >-
      Relative to the Core management hostname (e.g.
      https://core.example.com/api/v1/...)
info:
  title: AutoSecrets Core Management API
  version: v1
  description: |
    Session-authenticated management surface of the AutoSecrets Core control
    plane. This document is the contract: Core handlers are tested against it
    (kin-openapi response validation), the Web client's TypeScript types are
    generated from it (openapi-typescript), and CI lints it (redocly).

    The Agent surface (/agent/v1, mTLS + signed envelopes) is contracted by
    api/agent-envelope/envelope-v1.md and is intentionally not part of this
    document.

    Terminology follows CONTEXT.md (Application, Environment, Secret Bundle,
    Draft, Bundle Revision, Node Group, Assignment, Audit Event, ...).

    Error model: every error response carries {"error": string, "code": string}.
    Clients must switch on `code`, never on the human-readable `error` message.
    Codes: bad_request, unauthorized, forbidden, not_found, conflict, duplicate,
    internal, unavailable, payload_too_large.
security:
  - sessionCookie: []
...
paths:
...
  /api/v1/node-groups:
    get:
      operationId: listNodeGroups
      tags: [Fleet]
      summary: List Node Groups with members (cursor paginated)
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
        - $ref: "#/components/parameters/PageNumber"
      responses:
        "200":
          description: Node Groups newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeGroupPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createNodeGroup
      tags: [Fleet]
      summary: Create a Node Group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 64
      responses:
        "403":
          $ref: "#/components/responses/Forbidden"
        "201":
          description: Node Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRef"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Duplicate"
  /api/v1/node-groups/{groupID}:
    delete:
      operationId: deleteNodeGroup
      tags: [Fleet]
      summary: Delete a Node Group that has no active Assignments
      parameters:
        - $ref: "#/components/parameters/GroupID"
      responses:
        "204":
          description: Node Group deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/node-groups/{groupID}/nodes:
    post:
      operationId: addGroupMember
      tags: [Fleet]
      summary: Add a Managed Node to a Node Group
      parameters:
        - $ref: "#/components/parameters/GroupID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [node_id]
              properties:
                node_id:
                  type: string
      responses:
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "200":
          description: Member added
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/node-groups/{groupID}/nodes/{nodeID}:
    delete:
      operationId: removeGroupMember
      tags: [Fleet]
      summary: Remove a Managed Node from a Node Group
      parameters:
        - $ref: "#/components/parameters/GroupID"
        - $ref: "#/components/parameters/NodeID"
      responses:
        "403":
          $ref: "#/components/responses/Forbidden"
        "200":
          description: Member removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /api/v1/assignments:
    get:
      operationId: listAssignments
      tags: [Fleet]
      summary: List Assignments (cursor paginated)
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
        - $ref: "#/components/parameters/PageNumber"
      responses:
        "200":
          description: Assignments newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createAssignment
      tags: [Fleet]
      summary: >-
        Assign a Secret Bundle to a Node Group (follows the Desired Revision)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [group_id, application_id, environment_id]
              properties:
                group_id:
                  type: string
                application_id:
                  type: string
                environment_id:
                  type: string
                operation_reason:
                  $ref: "#/components/schemas/OperationReason"
      responses:
        "403":
          $ref: "#/components/responses/Forbidden"
        "201":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Duplicate"
  /api/v1/assignments/{assignmentID}:
    get:
      operationId: unassignmentState
      tags: [Fleet]
      summary: Assignment and its per-node Unassignment tasks
      parameters:
        - $ref: "#/components/parameters/AssignmentID"
      responses:
        "200":
          description: Assignment and cleanup tasks
          content:
            application/json:
              schema:
                type: object
                required: [assignment, tasks]
                properties:
                  assignment:
                    $ref: "#/components/schemas/Assignment"
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/UnassignmentTask"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/assignments/{assignmentID}/unassign:
    post:
      operationId: unassign
      tags: [Fleet]
      summary: Start persistent two-phase Unassignment for a Bundle Assignment
      parameters:
        - $ref: "#/components/parameters/AssignmentID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                operation_reason:
                  $ref: "#/components/schemas/OperationReason"
      responses:
        "202":
          description: Unassignment started with per-node tasks
          content:
            application/json:
              schema:
                type: object
                required: [id, status, tasks]
                properties:
                  id:
                    type: string
                  status:
                    type: string
                    enum: [removing]
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/UnassignmentTask"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/assignments/{assignmentID}/abandon-cleanup:
    post:
      operationId: abandonCleanup
      tags: [Fleet]
      summary: >-
        Stop waiting for unreachable nodes; records cleanup_unconfirmed per
        node
      parameters:
        - $ref: "#/components/parameters/AssignmentID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                operation_reason:
                  $ref: "#/components/schemas/OperationReason"
      responses:
        "200":
          description: >-
            Control-plane waiting ended with unconfirmed local cleanup
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [cleanup_unconfirmed]
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/v1/nodes:
    get:
      operationId: listNodes
      tags: [Fleet]
      summary: List Managed Nodes and their convergence state (cursor paginated)
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
        - $ref: "#/components/parameters/PageNumber"
      responses:
        "200":
          description: Nodes newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodePage"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/nodes/{nodeID}:
    patch:
      operationId: updateNode
      tags: [Fleet]
      summary: Adjust a Managed Node's settings, such as its polling interval
      parameters:
        - $ref: "#/components/parameters/NodeID"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [poll_interval_seconds]
              properties:
                poll_interval_seconds:
                  type: integer
                  minimum: 5
                  maximum: 86400
                  description: >-
                    How often the Agent polls for Desired State (5s-24h); the
                    Agent adopts it after its next heartbeat or desired poll.
      responses:
        "200":
          description: Updated node settings
          content:
            application/json:
              schema:
                type: object
                required: [id, poll_interval_seconds]
                properties:
                  id:
                    type: string
                  poll_interval_seconds:
                    type: integer
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/nodes/install-command:
    post:
      operationId: createInstallCommand
      tags: [Fleet]
      summary: >-
        Issue a ten-minute single-use Enrollment Token and Install Command
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 64
                  description: >-
                    Optional node name; a random one is generated when absent
                bundle_dir:
                  type: string
                  description: |
                    Optional override of the deployment root for the
                    Materialized Bundle on the node. Defaults to
                    ~/.autosecrets in the invoking user's home (the legacy
                    tool's layout); "~" expands to the sudo caller's home.
      responses:
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "201":
          description: Install Command; the Token appears in this response only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallCommand"
        "503":
          $ref: "#/components/responses/Unavailable"
...
components:
  parameters:
    PageLimit:
      name: limit
      in: query
      required: false
      description: Page size (default 25, max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
    PageCursor:
      name: cursor
      in: query
      required: false
      description: Opaque keyset position from the previous page
      schema:
        type: string
    PageNumber:
      name: page
      in: query
      required: false
      description: 1-based page number; used for jump-to-page
      schema:
        type: integer
        minimum: 1
...
    AssignmentID:
      name: assignmentID
      in: path
      required: true
      schema:
        type: string
      description: Assignment id (UUID)
...
    GroupID:
      name: groupID
      in: path
      required: true
      schema:
        type: string
      description: Node Group id (UUID)
    NodeID:
      name: nodeID
      in: path
      required: true
      schema:
        type: string
      description: Managed Node id (UUID)
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: autosecrets_session
      description: HttpOnly session cookie issued by /auth/login
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: >-
        Required on every non-GET request (double-submit token from
        /auth/login or /me)
  responses:
    BadRequest:
      description: Request payload or validation failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or expired session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Invalid CSRF token or forbidden action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Referenced resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: State conflict (e.g. stale Draft version)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Duplicate:
      description: A uniquely named resource already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unavailable:
      description: >-
        Dependency or configuration missing (e.g. install flow disabled)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    AdminRef:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
...
    OperationReason:
      type: object
      required: [category, explanation]
      properties:
        category:
          type: string
          enum:
            - maintenance
            - incident_response
            - access_change
            - configuration_correction
            - other
        explanation:
          type: string
          maxLength: 500
        external_ref:
          type: string
          maxLength: 128
    NodeGroup:
      type: object
      required: [id, name, member_ids]
      properties:
        id:
          type: string
        name:
          type: string
        member_ids:
          type: array
          items:
            type: string
    Assignment:
      type: object
      required:
        - id
        - group_id
        - group_name
        - application_id
        - environment_id
        - revision_id
        - status
        - created_at
      properties:
        id:
          type: string
        group_id:
          type: string
        group_name:
          type: string
        application_id:
          type: string
        environment_id:
          type: string
        revision_id:
          type: string
          description: >-
            Current Desired Revision, advanced by Publish and Rollback
        status:
          type: string
          enum: [active, removing]
        created_at:
          type: string
          format: date-time
    UnassignmentTask:
      type: object
      required: [assignment_id, node_id, status, updated_at]
      properties:
        assignment_id:
          type: string
        node_id:
          type: string
        status:
          type: string
          enum: [pending, cleaned, failed, offline, cleanup_unconfirmed]
        error:
          type: string
        updated_at:
          type: string
          format: date-time
          nullable: true
...
    Node:
      type: object
      required:
        - id
        - name
        - serial
        - created_at
        - last_seen_at
        - desired_etag
        - observed_revision
        - last_result
        - state
        - unassigned
        - poll_interval_seconds
      properties:
        id:
          type: string
        name:
          type: string
        serial:
          type: string
        created_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
          nullable: true
        desired_etag:
          type: string
        observed_revision:
          type: string
        last_result:
          type: string
        state:
          type: string
          enum: [never_online, healthy, converging, failed, offline]
          description: Derived primary state (ADR-0015)
        unassigned:
          type: boolean
        poll_interval_seconds:
          type: integer
          minimum: 5
          maximum: 86400
          description: >-
            How often the Agent polls for Desired State; Core delivers it in
            the desired/heartbeat responses and the Agent adopts it on its
            next pass.
...
    InstallCommand:
      type: object
      required: [command, expires_at]
      properties:
        command:
          type: string
          description: >-
            curl | sudo bash Install Command carrying the one-time Token
        expires_at:
          type: string
          format: date-time
...
    NodeGroupPage:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/NodeGroup"
        next_cursor:
          type: string
        total:
          type: integer
    AssignmentPage:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Assignment"
        next_cursor:
          type: string
        total:
          type: integer
    NodePage:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Node"
        next_cursor:
          type: string
        total:
          type: integer
...
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable message; never parse it
        code:
          type: string
          description: Stable machine-readable code; switch on this
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - duplicate
            - internal
            - unavailable
            - payload_too_large
            - mfa_enrollment_required
            - second_factor_required
            - challenge_expired
            - rate_limited